Move data classes from lib to app/models
