--- config/initializers/school_data_utilities.rb
@@ -1,3 +1,0 @@
-require 'csv'
-require Rails.root.join('lib', 'data', 'school_updater')
-require Rails.root.join('lib', 'data', 'school_mass_importer')
