+++ lib/tasks/data/manage_schools.rake
@@ -1,7 +1,3 @@
-require File.join(Rails.root, "lib", "data", "school_mass_importer")
-require File.join(Rails.root, "lib", "data", "school_manager")
-require File.join(Rails.root, "lib", "data", "school_updater")
-
 namespace :data do
   namespace :schools do
     desc "Import all GiaS (EduBase) data"
@@ -13,7 +9,7 @@ namespace :data do
         headers: true
       )
 
-      SchoolMassImporter.new(edubase_data, args[:email_override]).import
+      Bookings::Data::SchoolMassImporter.new(edubase_data, args[:email_override]).import
     end
 
     desc "Update schools"
@@ -23,16 +19,7 @@ namespace :data do
         headers: true
       )
 
-      SchoolUpdater.new(edubase_data).update
-    end
-
-    desc "Enhance school data using information captured in the questionnaire"
-    task :enhance, %i{questionnaire} => :environment do |_t, args|
-      response_data = CSV.parse(
-        File.read(args[:questionnaire]).scrub,
-        headers: true
-      )
-      SchoolEnhancer.new(response_data).enhance
+      Bookings::Data::SchoolUpdater.new(edubase_data).update
     end
 
     desc "Enable schools"
@@ -41,7 +28,7 @@ namespace :data do
         File.read(args[:urns]).scrub,
         headers: true
       )
-      SchoolManager.new(response_data).enable_urns
+      Bookings::Data::SchoolManager.new(response_data).enable_urns
     end
 
     desc "Disable schools"
@@ -50,7 +37,7 @@ namespace :data do
         File.read(args[:urns]).scrub,
         headers: true
       )
-      SchoolManager.new(response_data).disable_urns
+      Bookings::Data::SchoolManager.new(response_data).disable_urns
     end
   end
 end
